CDC-15 Refactor: sort.libs -> sort/{resource}.ts
