+++ apps/native/libs/sort/types.ts
@@ -1,0 +1,8 @@
+export type SortOptions = Record<string, SortOption>;
+
+export type SortOption = {
+  field: string;
+  orderBy: 'ASC' | 'DESC';
+  text: string;
+  authRequired?: boolean;
+};
